backend/requirements.txt
bcrypt==4.2.1
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask[async]==3.1.0
Flask-Caching==2.1.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
marshmallow==3.20.2
numpy>=1.26.0
pandas>=2.2.0
PyJWT==2.10.1
SQLAlchemy==2.0.38
typing_extensions==4.12.2
Werkzeug==3.1.3
scikit-learn>=1.3.0
scipy==1.11.4
transformers==4.38.2
torch==2.2.1
openai>=1.58.1
langchain>=0.1.9
langchain-openai>=0.3.7
langchain-core>=0.3.39
python-dotenv==1.0.1
asgiref>=3.7.2
